refactors all old fieldset to new, deletes old and adds tests to new
